+++ pom.xml
@@ -7,6 +7,7 @@
     <groupId>org.example</groupId>
     <artifactId>ARQUIVES_SO</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <packaging>jar</packaging>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
