--- src/main/java/org/example/compiler/lexer/TokenType.java
@@ -1,23 +1,0 @@
-package org.example.compiler.lexer;
-
-public enum TokenType {
-    SLASH,
-    SEMICOLON,
-    DOT,
-    TWODOTS,
-    PATH,
-    INDENTIFIER,
-    SPECIAL,
-    WHITESPACE,
-    ARGUMENTS,
-    MKDIR,
-    RM,
-    TOUCH,
-    CD,
-    LS,
-    PWD,
-    NUMBER,
-    IDENTIFIER,
-    NANO,
-    EOF,
-}
