--- src/main/java/org/example/compiler/parser/command/CommandNode.java
@@ -1,11 +1,0 @@
-package org.example.compiler.parser.command;
-
-import org.example.system.FileSystem;
-
-import java.io.IOException;
-
-public abstract class CommandNode {
-
-    public abstract String execute(FileSystem context) throws IOException;
-
-}
