--- src/main/java/org/example/compiler/pipeline/component/IOComponent.java
@@ -1,5 +1,0 @@
-package org.example.compiler.pipeline.component;
-
-public class IOComponent<T extends  IOComponent<T>> {
-
-}
