--- src/main/java/org/example/compiler/pipeline/execptions/UnexpectInputType.java
@@ -1,5 +1,0 @@
-package org.example.compiler.pipeline.execptions;
-
-public class UnexpectInputType extends Throwable {
-
-}
