+++ src/main/java/org/example/console/lexer/TokenType.java
@@ -1,0 +1,23 @@
+package org.example.console.lexer;
+
+public enum TokenType {
+    SLASH,
+    SEMICOLON,
+    DOT,
+    TWODOTS,
+    PATH,
+    INDENTIFIER,
+    SPECIAL,
+    WHITESPACE,
+    ARGUMENTS,
+    MKDIR,
+    RM,
+    TOUCH,
+    CD,
+    LS,
+    PWD,
+    NUMBER,
+    IDENTIFIER,
+    NANO,
+    EOF,
+}
