+++ src/main/java/org/example/console/parser/command/CommandNode.java
@@ -1,0 +1,11 @@
+package org.example.console.parser.command;
+
+import org.example.system.FileSystem;
+
+import java.io.IOException;
+
+public abstract class CommandNode {
+
+    public abstract String execute(FileSystem context) throws IOException;
+
+}
