+++ src/main/java/org/example/console/pipeline/component/IOComponent.java
@@ -1,0 +1,5 @@
+package org.example.console.pipeline.component;
+
+public class IOComponent<T extends  IOComponent<T>> {
+
+}
