+++ src/main/java/org/example/console/pipeline/execptions/UnexpectInputType.java
@@ -1,0 +1,5 @@
+package org.example.console.pipeline.execptions;
+
+public class UnexpectInputType extends Throwable {
+
+}
