+++ src/main/java/org/example/console/util/CMDInput.java
@@ -1,0 +1,15 @@
+package org.example.console.util;
+
+import org.example.console.pipeline.component.IOComponent;
+
+public class CMDInput extends IOComponent<CMDInput> {
+    private final String command;
+
+    public CMDInput(String command) {
+        this.command = command;
+    }
+
+    public String getCommand() {
+        return command;
+    }
+}
