+++ src/main/java/org/example/gui/ArquiveCatcher.java
@@ -1,6 +1,6 @@
 package org.example.gui;
 
-import org.example.compiler.pipeline.component.IOComponent;
+import org.example.console.pipeline.component.IOComponent;
 import org.example.system.arquives.Arquive;
 
 public class ArquiveCatcher extends IOComponent<ArquiveCatcher> {
