+++ src/main/resources/cmd.test
@@ -1,0 +1,1 @@
+ls
